[FIX] demo_mn_console: fix compiler warning
On 32-bit machines the compiler says:
demo_main.c:460:5: note: expected 'char *' but argument is of type 'unsigned char *'
demo_main.c:530:5: warning: format '%ld' expects argument of type 'long int', but argument 2 has type 'unsigned int' [-Wformat]
demo_main.c:531:5: warning: format '%ld' expects argument of type 'long int', but argument 2 has type 'unsigned int' [-Wformat]

On 64-bit machines the data type of sizeof() is unsigned long. So we need
to cast the result of sizeof() to unsigned long and use %lu.

+++ Examples/X86/Generic/demo_mn_console/demo_main.c
@@ -527,8 +527,8 @@ int  main (int argc, char **argv)
 #endif
 
     printf("Initializing process image...\n");
-    printf("Size of input process image: %ld\n", sizeof(AppProcessImageIn_g));
-    printf("Size of output process image: %ld\n", sizeof (AppProcessImageOut_g));
+    printf("Size of input process image: %lu\n", (unsigned long) sizeof(AppProcessImageIn_g));
+    printf("Size of output process image: %lu\n", (unsigned long) sizeof (AppProcessImageOut_g));
     AppProcessImageCopyJob_g.m_fNonBlocking = FALSE;
     AppProcessImageCopyJob_g.m_uiPriority = 0;
     AppProcessImageCopyJob_g.m_In.m_pPart = &AppProcessImageIn_g;
